work on training

+++ segmentation/src/utils.py
@@ -38,7 +38,7 @@ class CrossEntropyLoss2d(nn.Module):
             targets_m = targets.clone()
             targets_m -= 1
             # loss_all = self.ce_loss(inputs, targets_m.long())
-            loss_all = self.ce_loss(inputs.long(), targets_m.long())
+            loss_all = self.ce_loss(inputs.float(), targets_m.long())
             # loss_all = torch.nn.functional.cross_entropy(inputs, targets_m)
             number_of_pixels_per_class = \
                 torch.bincount(targets.flatten().type(self.dtype),
